moved Util class to viewmodel package and rename to ViewModelUtil
